mcu: Try to parse down-sampled images that have one component.
I'm really not sure on the intended behaviour here, but since all jpeg decoders I know parse this, I'm going to  assume that I'm the wrong one.

We can't have an image with a single component being down-sampled, but apparently they exist, so this commits allows the decoder to actually decode that.

Fixes https://github.com/etemesi254/zune-jpeg/issues/13